requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bayesian-optimization==3.1.0
beautifulsoup4==4.14.0
bidict==0.23.1
bleak==0.22.3
bleak-winrt==1.2.0
blinker==1.8.2
boxmot==15.0.2
branca==0.7.1
brotli==1.2.0
certifi==2025.8.3
cffi==1.16.0
charset-normalizer==3.4.3
click==8.3.0
cmake==3.28.3
colorama==0.4.6
comm==0.2.2
comtypes==1.4.1
ConfigArgParse==1.7.1
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.7
dlib==19.24.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.135.2
fastjsonschema==2.21.1
ffmpy==1.0.0
filelock==3.19.1
filterpy==1.4.5
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.6
flatbuffers==25.2.10
folium==0.16.0
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
futures==3.0.5
gast==0.6.0
gdown==5.2.0
geopandas==1.1.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grad-cam==1.5.5
gradio==6.10.0
gradio_client==2.4.0
groovy==0.1.2
grpcio==1.71.0
gTTS==2.5.1
h11==0.16.0
h5py==3.13.0
hf-gradio==0.3.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
importlib_metadata==7.1.0
inference-sdk==0.54.2
install==1.3.5
ipython==8.37.0
ipywidgets==8.1.7
iso8601==2.1.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kagglehub==1.0.0
kagglesdk==0.1.16
keras==3.9.2
kiwisolver==1.4.9
lap==0.5.12
lapx==0.5.12
lazy_loader==0.4
libclang==18.1.1
loguru==0.7.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.9
ml_dtypes==0.5.1
mplcursors==0.7
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
mysql-connector-python==9.0.0
namex==0.0.9
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
obd==0.7.2
open3d==0.19.0
opencv-contrib-python==4.9.0.80
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.15.0
orjson==3.11.7
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
parso==0.8.4
pillow==11.3.0
Pint==0.20.1
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.3
psutil==5.9.8
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pygame==2.5.2
Pygments==2.19.1
PyMySQL==1.1.1
pyogrio==0.11.0
PyOpenGL==3.1.7
pyparsing==3.2.5
pypiwin32==223
pyproj==3.7.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
PySocks==1.7.1
pyspin==1.1.1
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-multipart==0.0.22
python-socketio==5.11.2
pyttsx3==2.90
pytz==2025.2
PyWavelets==1.8.0
pywin32==306
PyYAML==6.0.3
QtPy==2.4.1
qtwidgets==1.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
retrying==1.3.4
rich==14.0.0
rpds-py==0.25.1
safehttpx==0.1.7
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
selenium==4.28.1
semantic-version==2.10.0
shapely==2.1.1
shellingham==1.5.4
simple-websocket==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.8
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.52.1
supervision==0.26.1
sympy==1.14.0
tenacity==8.3.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
ttach==0.0.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.149
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.42.0
virtualenv==20.26.3
wcwidth==0.2.14
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.14
win32_setctime==1.2.0
wrapt==1.17.2
wsproto==1.2.0
xyzservices==2023.10.1
yacs==0.1.8
yarl==1.20.1
zipp==3.19.1
